refactor: move receipt and transaction builders to utils folder
