+++ src/services/filter/filters/mod.rs
@@ -9,7 +9,6 @@
 pub mod evm {
 	pub mod filter;
 	pub mod helpers;
-	pub mod test_helpers;
 }
 pub mod stellar {
 	pub mod filter;
